Time: 4 ms (46.5%), Space: 7 MB (61.55%) - LeetHub

+++ 0633-sum-of-square-numbers/0633-sum-of-square-numbers.cpp
@@ -1,8 +1,8 @@
 class Solution {
 public:
     bool judgeSquareSum(int c) {
-        long long st=0;
-        long long en=sqrt(c);
+        long st=0;
+        long en=sqrt(c);
 
         while(st<=en)
         {
